Move spark.executor.ExecutorLogUrlHandler to spark.LogUrlHandler
